+++ src/src/DataStructs/NodesHandler.cpp
@@ -552,13 +552,13 @@ NodesMap::const_iterator NodesHandler::find(uint8_t unit_nr) const
   return _nodes.find(unit_nr);
 }
 
-bool NodesHandler::refreshNodeList(unsigned long max_age_allowed, unsigned long& max_age)
+bool NodesHandler::refreshNodeList(uint32_t max_age_allowed, uint32_t& max_age)
 {
   max_age = 0;
   bool nodeRemoved = false;
 
   for (auto it = _nodes.begin(); it != _nodes.end();) {
-    unsigned long age = it->second.getAge();
+    uint32_t age = it->second.getAge();
     if (age > max_age_allowed) {
       bool mustErase = true;
       #ifdef USES_ESPEASY_NOW
